Create an actual interface rather than interface{}
Problem with interface{} is that we need all the objects to be in their
"unstructured" form, and this won't work for built-in types that will
still have to be converted to that form. We can further improve the
allocations by removing the need for interface{}, and built a real
interface instead that can be implemented for built-in types.